Switch boot C code over to new args handling
Since a bunch of these are going to be added in order to implement the
boot dispatchers, we may as well write them as they eventually should
look rather than having more things to tweak later. This also gives an
opportunity to get a few more missing pieces in place with regards to
the new args handling: the identity map and support for fetching named
arguments.

We do need a fudge to map between the previous arg passing support and
the new one. And very fudgey it is indeed. This won't be needed for the
far more common case of running bytecode; we'll just provide parallel
frame invoke functions there.