Re-work uninlining for the new callstack scheme
Overall, this is a simplification. Since we don't have to care about
uninlining in the middle of a callstack any more, the code to deal with
clearly disappears. However, a further benefit we get is that uninline
can do its work, and then leave its caller to do the appropriate kind of
next action on the stack top, whether it's the deopt one case or lazy
deopt all case. Thus it can work without knowing what kind of deopt we
are doing.

Furthermore, we can now deopt into frames on the stack, not on the heap,
which reduces the cost of the deopt and doesn't force heap allocations
further down.

Unfortunately, while this gets us through the NQP build and test,
there's an explosion in the Rakudo core setting build that will still
need to be tracked down.