Stub in the boot-syscall mechanism
This provides a way to register such calls, and sets them up at startup.
A call has an unchecked implementation body that will pull received args
out without needing to validate them, along with metadata used to do the
validation. A boot-syscall, when it's fully implemented, will on its
first dispatch do these checks and set up the guards. The ultimate goal
is that we can do standard guard elimination on these in spesh, then
call the C function, providing a far more uniform interface for the VM
interanls.